All CLI commands in one module
