+++ crates/cli/src/commands.rs
@@ -1,3 +1,3 @@
 pub mod auth;
 pub mod check_pr;
-pub mod config;
+pub mod config_cmd;
